+++ src/read_ab_initio.f90
@@ -7,7 +7,7 @@ implicit none
 
 integer(i4b) :: i, j, k, ii, img_num
 character(20) :: dummy
-real(sp) :: vec(3), qgbohr(nx)
+real(dp) :: vec(3), qgbohr(nx)
 
 !Loop over all images
 do img_num=1,nimg
